mathtrainer/pyproject.toml
[tool.poetry]
name = "mathtrainer"
version = "0.1.0"
description = "peruskoulumatematiikan harjoittelusovellus"
authors = ["Ari <ari.virtanen@tuni.fi>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
invoke = "^1.7.3"
prettytable = "^3.5.0"
python-dotenv = "^0.21.0"


[tool.poetry.group.dev.dependencies]
pytest = "^7.2.0"
coverage = "^6.5.0"
pylint = "^2.15.6"
autopep8 = "^2.0.0"
pytest-dotenv = "^0.5.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
